backend-python/requirements.txt
whisperx
torch
numpy
fastapi
uvicorn
python-multipart
soundfile
librosa
langid
